requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-forms==2.0
dnspython==2.4.2
environs==9.5.0
flower==2.0.1
frozenlist==1.4.0
hiredis==2.2.3
humanize==4.8.0
idna==3.4
kombu==5.3.2
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
gunicorn==21.2.0


